refactor(resume-validation): split session_on_disk tests into sibling file to honor 1K-line cap
Moved the eight session_on_disk test functions (and their helpers SCAN_COUNTER,
scan_temp_home, running_as_root) from amplifier_stub.rs (1049 → 906 lines) into
a new sibling module amplifier_stub_scan_tests.rs (148 lines), wired up via
#[path] attribute per the pane_ledger_tests.rs pattern.

All 8 tests verified running and passing:
  - session_on_disk_present_under_cwd_slug
  - session_on_disk_present_under_divergent_slug
  - session_on_disk_absent_when_store_readable
  - session_on_disk_absent_when_projects_dir_missing
  - session_on_disk_unreadable_projects_dir_fails_open
  - session_on_disk_unreadable_project_subdir_fails_open
  - session_on_disk_listable_not_traversable_projects_fails_open
  - session_on_disk_matches_a_file_only_as_absent

Clippy: clean (no warnings).
